Graph arreglar 6merge 21/11

+++ src/main/java/com/example/atlantis/model/GraphqlInput.java
@@ -57,8 +57,8 @@ public class GraphqlInput {
         private Integer num_estrellas;
         private TipoHotel tipo_hotel;
         private String telefono;
-        private Integer latitud;
-        private Integer longitud;
+        private Double latitud;
+        private Double longitud;
         private LoginInput email;
         private String url_icono;
         private String url_imagen_general;
@@ -77,8 +77,8 @@ public class GraphqlInput {
         private Integer num_estrellas;
         private TipoHotel tipoHotel;
         private String telefono;
-        private Integer latitud;
-        private Integer longitud;
+        private Double latitud;
+        private Double longitud;
         private LoginInput email;
         private String url_icono;
         private String url_imagen_general;
